[CPDLP-3555] Move valid test data generators to services folder
