+++ app/services/valid_test_data_generators/completed_mentor_generator.rb
@@ -1,26 +1,29 @@
 # frozen_string_literal: true
 
-require "tasks/school_urn_generator"
-require "tasks/trn_generator"
 require "active_support/testing/time_helpers"
 
-module ValidTestDataGenerator
-  class LeadProviderPopulater
+module ValidTestDataGenerators
+  class CompletedMentorGenerator
     include ActiveSupport::Testing::TimeHelpers
 
     class << self
-      def call(name:, total_schools: 10, participants_per_school: 50, cohort: Cohort.current)
-        new(name:, cohort:).call(total_schools:, participants_per_school:)
+      def call(name:, cohort: Cohort.current, total_completed_mentors: 30)
+        new(name:, cohort:).call(total_completed_mentors:)
       end
     end
 
-    def call(total_schools:, participants_per_school:)
-      generate_new_schools(count: total_schools)
+    def call(total_completed_mentors:)
+      school = create_fip_school_with_cohort(urn: Helpers::SchoolUrnGenerator.next)
 
-      lead_provider.schools.order("created_at desc").limit(total_schools).each do |school|
-        sparsity_uplift = weighted_choice(selection: [true, false], odds: [11, 89])
-        pupil_premium_uplift = weighted_choice(selection: [true, false], odds: [11, 39])
-        find_or_create_participants(school:, number_of_participants: participants_per_school, sparsity_uplift:, pupil_premium_uplift:)
+      sparsity_uplift = weighted_choice(selection: [true, false], odds: [11, 89])
+      pupil_premium_uplift = weighted_choice(selection: [true, false], odds: [11, 39])
+
+      total_completed_mentors.times do
+        mentor = create_mentor(school:, sparsity_uplift:, pupil_premium_uplift:)
+
+        completion_date = rand(10..100).days.from_now.to_date
+        completion_reason = ParticipantProfile::Mentor.mentor_completion_reasons.values.sample
+        mentor.complete_training!(completion_date:, completion_reason:)
       end
     end
 
@@ -29,49 +32,29 @@ module ValidTestDataGenerator
     attr_reader :lead_provider, :cohort
 
     def initialize(name:, cohort:)
-      @lead_provider = ::LeadProvider.find_or_create_by!(name:)
+      @lead_provider = ::LeadProvider.find_by!(name:)
       @cohort = cohort
     end
 
-    def generate_new_schools(count:)
-      count.times { create_fip_school_with_cohort(urn: SchoolURNGenerator.next) }
-    end
+    def create_mentor(school:, sparsity_uplift:, pupil_premium_uplift:)
+      status = weighted_choice(selection: %w[active withdrawn], odds: [6, 1])
 
-    def find_or_create_participants(school:, number_of_participants:, sparsity_uplift:, pupil_premium_uplift:)
-      generate_new_participants(school:, count: number_of_participants - school.ecf_participants.count, sparsity_uplift:, pupil_premium_uplift:) if school.ecf_participants.count < number_of_participants
-    end
-
-    def generate_new_participants(school:, count:, sparsity_uplift:, pupil_premium_uplift:)
-      while count.positive?
-        status = weighted_choice(selection: %w[active withdrawn], odds: [6, 1])
-        profile_type = weighted_choice(selection: %i[mentor ect], odds: [9, 1])
-        participant_identity = create_random_participant_identity
-
-        count -= 1
-        if profile_type == :mentor
-          mentor = create_participant(participant_identity:, school_cohort: school_cohort(school:), profile_type: :mentor, status:, sparsity_uplift:, pupil_premium_uplift:)
-          rand(0..3).times do
-            create_participant(participant_identity:, school_cohort: school_cohort(school:), profile_type: :ect, mentor_profile: mentor, status:, sparsity_uplift:, pupil_premium_uplift:)
-            count -= 1
-          end
-        else
-          create_participant(participant_identity:, school_cohort: school_cohort(school:), profile_type: :ect, status:, sparsity_uplift:, pupil_premium_uplift:)
-        end
+      mentor = create_participant(school_cohort: school_cohort(school:), profile_type: :mentor, status:, sparsity_uplift:, pupil_premium_uplift:)
+      rand(0..3).times do
+        create_participant(school_cohort: school_cohort(school:), profile_type: :ect, mentor_profile: mentor, status:, sparsity_uplift:, pupil_premium_uplift:)
       end
+
+      mentor
     end
 
-    def create_random_participant_identity
+    def create_participant(school_cohort:, profile_type: :ect, mentor_profile: nil, status: "active", sparsity_uplift: false, pupil_premium_uplift: false)
       name = Faker::Name.name
       user = User.create!(full_name: name, email: Faker::Internet.email(name:))
-      TeacherProfile.create!(user:, trn: random_or_nil_trn)
-      Identity::Create.call(user:, origin: :ecf)
-    end
-
-    def create_participant(participant_identity:, school_cohort:, profile_type: :ect, mentor_profile: nil, status: "active", sparsity_uplift: false, pupil_premium_uplift: false)
-      user = participant_identity.user
-      teacher_profile = user.teacher_profile
+      teacher_profile = TeacherProfile.create!(user:, trn: random_or_nil_trn)
       schedule = ecf_schedules.sample
+      participant_identity = Identity::Create.call(user:, origin: :ecf)
 
+      lead_provider = school_cohort.lead_provider
       cpd_lead_provider = lead_provider.cpd_lead_provider
       statement = Finance::Statement::ECF.where(
         cpd_lead_provider:,
@@ -100,7 +83,7 @@ module ValidTestDataGenerator
 
         Induction::Enrol.call(participant_profile: profile, induction_programme:)
 
-        return unless profile.active_record?
+        return profile unless profile.active_record?
 
         started_declaration = travel_to profile.schedule.milestones.first.start_date + rand(5.days).seconds do
           RecordDeclaration.new(
@@ -111,9 +94,9 @@ module ValidTestDataGenerator
             declaration_type: "started",
           ).call
         end
-        return unless started_declaration
+        return profile unless started_declaration
 
-        return if profile.schedule.milestones.second.start_date > Date.current
+        return profile if profile.schedule.milestones.second.start_date > Date.current
 
         started_declaration.make_payable!
         started_declaration.update!(
@@ -129,7 +112,7 @@ module ValidTestDataGenerator
           state: started_declaration.state,
         )
 
-        return if (profile.schedule.milestones.second.start_date + 1.day) > Time.zone.now
+        return profile if (profile.schedule.milestones.second.start_date + 1.day) > Time.zone.now
 
         RecordDeclaration.new(
           participant_id: user.tap(&:reload).id,
@@ -160,9 +143,9 @@ module ValidTestDataGenerator
             declaration_type: "started",
           ).call
         end
-        return unless started_declaration
+        return profile unless started_declaration
 
-        return if profile.schedule.milestones.second.start_date > Date.current
+        return profile if profile.schedule.milestones.second.start_date > Date.current
 
         started_declaration.make_payable!
         started_declaration.update!(
@@ -178,7 +161,7 @@ module ValidTestDataGenerator
           state: started_declaration.state,
         )
 
-        return if (profile.schedule.milestones.second.start_date + 1.day) > Time.zone.now
+        return profile if (profile.schedule.milestones.second.start_date + 1.day) > Time.zone.now
 
         RecordDeclaration.new(
           participant_id: user.tap(&:reload).id,
@@ -219,6 +202,7 @@ module ValidTestDataGenerator
         training_programme: "full_induction_programme",
         school_cohort:,
       )
+      school
     end
 
     def attach_partnership_to_school(school:)
@@ -237,57 +221,13 @@ module ValidTestDataGenerator
     end
 
     def random_or_nil_trn
-      [true, false].sample ? nil : TRNGenerator.next
+      [true, false].sample ? nil : Helpers::TrnGenerator.next
     end
 
     def weighted_choice(selection:, odds:)
       selection.each_with_index.map { |item, index|
         [item] * odds[index]
       }.flatten.sample
-    end
-  end
-
-  class AmbitionSpecificPopulater < LeadProviderPopulater
-    class << self
-      FIRST_AMBITION_SEED_DATA_TIME = ("2022-08-18 13:43".."2022-08-18 13:49")
-
-      def call(name:, total_schools: 3, participants_per_school: 3000, cohort: Cohort.current)
-        generator = new(name:, cohort:)
-
-        generator.remove_old_data(created_at: FIRST_AMBITION_SEED_DATA_TIME)
-        generator.call(total_schools:, participants_per_school:)
-      end
-    end
-
-    def remove_old_data(created_at:)
-      lead_provider.ecf_participants.where(created_at:).find_each(&:destroy)
-      schools = lead_provider.schools.where(created_at:)
-      schools.each do |school|
-        partnership = Partnership.find_by(lead_provider:, school:, cohort:)
-        delivery_partner = partnership.delivery_partner
-        provider_relationship = ProviderRelationship.find_by(lead_provider:,
-                                                             cohort:,
-                                                             delivery_partner:)
-        provider_relationship.destroy!
-        partnership.destroy!
-        delivery_partner.destroy!
-        school_cohort = SchoolCohort.find_by(school:, cohort:, induction_programme_choice: "full_induction_programme")
-        school_cohort.ecf_participant_profiles.destroy_all
-        school_cohort.destroy!
-        school.destroy!
-      end
-    end
-
-    def generate_new_participants(school:, count:, sparsity_uplift:, pupil_premium_uplift:)
-      (count / 2).times do
-        status = "active"
-        mentor = create_participant(participant_identity: create_random_participant_identity, school_cohort: school_cohort(school:), profile_type: :mentor, status:, sparsity_uplift:, pupil_premium_uplift:)
-        create_participant(participant_identity: create_random_participant_identity, school_cohort: school_cohort(school:), profile_type: :ect, mentor_profile: mentor, status:, sparsity_uplift:, pupil_premium_uplift:)
-      end
-    end
-
-    def random_or_nil_trn
-      TRNGenerator.next
     end
   end
 end
