+++ app/services/valid_test_data_generators/pending_npq_applications_populater.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
-require "tasks/valid_test_data_generator/base_populater"
-
-module ValidTestDataGenerator
+module ValidTestDataGenerators
   class PendingNPQApplicationsPopulater < BasePopulater
     include ActiveSupport::Testing::TimeHelpers
