+++ db/new_seeds/scenarios/participants/ects/ect_after_cohort_transfer.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "active_support/testing/time_helpers"
+
 module NewSeeds
   module Scenarios
     module Participants
